Add Alertmanager controller (#201)
Reconciles the Alertmanager secret created by the observability-operator Helm chart and load the configuration and templates from the secret into Mimir Alertmanager.
It also watches the Mimir Alertmanager pod and re-queue events to the controller when the pod is restarted, so the configuration is reloaded.

Predicates are being used to filter the secret and pods being watched to only act on Alertmanager related resources.
Finalizers are not used by this controller as the configuration is never deleted, only updated.

---------

Co-authored-by: Taylor Bot <dev@giantswarm.io>
Co-authored-by: Quentin Bisson <quentin@giantswarm.io>